engine: extract esm/grids.js
