+++ engine/esm/index.js
@@ -213,6 +213,7 @@ export { SpaceTimeController } from "./space_time_controller.js";
 export { KeplerianElements, BodyAngles, Planets } from "./planets.js";
 export { Place } from "./place.js";
 export { FolderUp } from "./folder_up.js";
+export { Grids } from "./grids.js";
 
 export {
     PointType,
@@ -249,7 +250,6 @@ export {
     Pointing,
     Folder,
     FolderBrowser,
-    Grids,
     ViewMoverKenBurnsStyle,
     KeplerVertex,
     Layer,
